rename eval to exec
