+++ core/exec/src/mill/exec/Cached.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 case class Cached(value: ujson.Value, valueHash: Int, inputsHash: Int)
