+++ core/exec/src/mill/exec/EvalResults.scala
@@ -1,4 +1,5 @@
-package mill.eval
+package mill.exec
+
 import mill.api.*
 import mill.api.Strict.Agg
 import mill.define.*
