+++ core/exec/src/mill/exec/EvaluatorLogs.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.define.{InputImpl, Task}
 import mill.main.client.OutFiles
