+++ core/exec/src/mill/exec/EvaluatorPathsResolver.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.define.{NamedTask, Segments}
