+++ core/exec/src/mill/exec/ThreadNumberer.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 /**
  * Small class to take named threads and assign them stable integer IDs
