+++ core/exec/test/resources/examples/javac/resources/META-INF/License
