+++ core/exec/test/resources/examples/javac/resources/hello.txt
