+++ core/exec/test/resources/examples/javac/src/Foo.java
