+++ scalalib/test/src/mill/scalalib/HelloWorldTests.scala
@@ -7,7 +7,7 @@ import scala.util.Using
 import mill.*
 import mill.api.Result
 import mill.define.Discover
-import mill.eval.EvaluatorPaths
+import mill.exec.EvaluatorPaths
 import mill.testkit.UnitTester
 import mill.testkit.TestBaseModule
 import utest.*
