+++ testkit/package.mill
@@ -3,7 +3,7 @@ package build.testkit
 import mill._
 
 object `package` extends RootModule with build.MillPublishScalaModule {
-  def moduleDeps = Seq(build.core.eval, build.main.util, build.main)
+  def moduleDeps = Seq(build.core, build.main.util, build.main)
   def ivyDeps = Agg(build.Deps.TestDeps.utest)
 
   def sources =
